refactor: refactor based on review
